requirements.txt
alembic==1.4.2
APScheduler==3.6.3
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.1
